+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added 
 
 - [#335](https://github.com/XenitAB/spegel/pull/335) Add k3s to compatibility guide.
+- [#339](https://github.com/XenitAB/spegel/pull/339) Extend OCI client tests.
 
 ### Changed